requirements.txt
fastapi
python-dotenv
pytz
azure-identity
azure-keyvault-secrets
azure-storage-blob
psycopg2
uuid
uvicorn
python-multipart
